Only trreat granular results as ranges in apply
